Fixing <DropdownMenu> tag selection during editing.
Tags that were already there wouldn't unselect if their checkboxes
were clicked. Only way to unselect was via the All button or by
clicking the (x) on the tag inside the input field when the menu
was closed.

Also changed the initial setting of selectedOptions to be done via
useEffect for better clarity.