<DropdownMenu> label "double click" fix.
Turns out that having a <label> attached to an <input> via 'htmlFor'
makes so that clicking the label triggers the onClick for the input.
This was causing a double-click to happen when the label was clicked,
because of the onClick on the parent <div> plus the propagated from the
label, making the checkbox "not" check.