Fixing emptiness animation duration inconsistency.
The problem was because the duration was being set for all animations
based on the duration of the scroll. Now, it uses a CSS variable to set
the duration of the fadeText animation.

Because of this, the useEffect that cleared the animation duration was
also removed and the fadeText duration was bumped to 2 seconds.