Fixing a maximum-depth error on <Tag>.
The error was caused by an improper jury-rig to fix an error caused by a
re-render after returning from <DeleteAllMessage>. The jury-rig worked
by repeating the useEffect until all elements finished loading so their
correct sizes could be used for the '--too-big' calculation, but this
was only tested when there were only a few <Tag>s. When more were added,
it caused overhead which led things to maximum-depth.

The new fix works by ditching the jury-rig in favour of using an
Observer to watch the parentDiv and rerun the useEffect based on that.
Also, this also works for the issue of improper '--too-big' on first
loading the page, which removes the necessity to run the useEffect on
'scrollingText.current.offsetWidth'.