<EditTag> "New tag" button.
The Route, Controller and api call for creating a new tag now properly
return the created tag;

The 'actuallyEmpty' useState was merged with tags to prevent an error
where, when an empty tag was added, one could happen to finish updating
before the other, leading to an incorrect state being displayed. The
useEffect that updated 'setActuallyEmpty' every time tags changed was
done away in favour of just updating the tags.empty property inside
'updateTagLabel()' alongside the label value.