Fixing .epub cover image not loading properly.
The previous approach would not work on all epubs, due to how images may be
inserted into the xthml file. This new solution assumes that the epub was
made using OEBPS, which streamlines the file placement inside it.

Also done away with awaiting the Promisse to load the cover, as, for some
reason, the Promisse was returning null after the changes to how the cover
is searched inside the file.