<NavBar> refactor to hold <SideMenu>.
I mean, it makes sense. Both need the same props and those props are
only used by the <NavBar>. Also, since the <HamburgerMenu> is part of
<NavBar> and it's used to open the <SideMenu>, it kinda makes sense for
that to be part of the navigation bar aswell.

Also makes the rest of the code cleaner, since it was all the same on
all pages.