Proper separation of responsabilities.
Previously, the front-end was organizing and doing individual calls to
save a new book on the database. It worked, but made things a bit spaghetti
and also unoptimized (which would cause problems if the front and back-end
weren't on the same machine).

Now, the front-end sends the entire 'book' object as is and recieves any
'book' object as is, leaving the back-end to figure everything out.