Incorrect revert icon while editing fix.
The state variable 'originalCover' only gets set after going through the
file selection process, which doesn't happen when editing a book. This
would cause the revert (to previous cover image) button to appear by
default when first going to edit a book with a cover, as the
'originalCover' (null) would differ from the 'book.cover'.