Another author deletion fix.
When a book is edited and saved, the authors' list will be checked to
see if any of the previous authors have gotten deleted. If they're
deleted (from the book), they get checked to see if any other book
still have them. If no more book has that author, it gets deleted
from the database.