Adding 'useScrollable' to <TitleContainer>.
Had to change the <h1>'s structure inside <TitleContainer>, because
the way it was before would make so that 'useScrollable' wouldn't get a
textRef when it was defined, leading to it always being false. Now,
either way the <h1> will always get the useRef.

This took me way longer for me to notice why it was broken than I'd like
to admit, but at least it looks cleaner now.