Splitting <BookCard>'s publisher and release strings.
Due to difference in sizing, in Portuguese, in most cases the release
date was being pushed to a newline already, which made it weird when one
had a separation between the underlined text and regular text while the
other doesn't.

Like this, not only does it make things clear by always keeping underlined
alone on top, but also keeps the visuals consistent between languages.