Changing epub-reader light color scheme.
Previously, it had been made to match the aesthetic of the light mode
for the rest of the app. The problem is that not only is the dark mode
for the reader not 1:1 with the dark mode for the rest of the app, but
also, the light mode for the reader looked very out-of-place with the
dark mode navbar (when the app itself was in dark mode).

Now, to align it with the reader dark mode, its light mode looks more
neutral, as to blend in with both color modes (for the app itself).